+++ Documentation/ChangeLog.md
@@ -11,7 +11,10 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Changed
 
-* Player and Monster movement is now velcity based.
+* Player and Monster movement is now velocity based.
+
+### Fixed
+